SW-4021 Move scheduled observations notification to the observations root page
- was previously only included in the list/map views
- move it up so it shows for the empty state as well
