+++ src/components/PlantingSites/PlantingSiteView.tsx
@@ -131,8 +131,10 @@ export default function PlantingSiteView(): JSX.Element {
           </Grid>
         )}
         {plantingSite?.boundary && !plantingSite.plantingZones && (
-          <Grid item xs={12}>
-            <SimplePlantingSite plantingSite={plantingSite} />
+          <Grid container flexGrow={1}>
+            <Grid item xs={12} display='flex'>
+              <SimplePlantingSite plantingSite={plantingSite} />
+            </Grid>
           </Grid>
         )}
       </Card>
